Added image data generation to Mock Instrument class (#552)
Create methods to return single channel mock image data with specified width, height, and bit depth
Maintenance around Mocks/fakes:
Updated Mock to inherit from FakeInstrument, rename to SwissArmyFake, put new functionality there.
Removed SwissArmyFake and FakeInstrument from the Instrument namespace.
Remove __getitem__ from SwissArmyFake. This was originally introduced (in a suspected nonfunctional
manner) in 3c5c158470e5744af9ff881c27e050ba5feb006a, but its current purpose is unknown.
