+++ pymeasure/instruments/__init__.py
@@ -24,7 +24,6 @@
 
 from ..errors import RangeError, RangeException
 from .instrument import Instrument
-from .mock import Mock
 from .resources import list_resources
 from .validators import discreteTruncate
