+++ pymeasure/instruments/instrument.py
@@ -23,11 +23,7 @@
 #
 
 import logging
-import re
-
 import numpy as np
-
-from pymeasure.adapters import FakeAdapter
 from pymeasure.adapters.visa import VISAAdapter
 
 log = logging.getLogger(__name__)
@@ -527,52 +523,3 @@ class Instrument:
             return errors
         else:
             raise NotImplementedError("Non SCPI instruments require implementation in subclasses")
-
-
-class FakeInstrument(Instrument):
-    """ Provides a fake implementation of the Instrument class
-    for testing purposes.
-    """
-
-    def __init__(self, adapter=None, name=None, includeSCPI=False, **kwargs):
-        super().__init__(
-            FakeAdapter(**kwargs),
-            name or "Fake Instrument",
-            includeSCPI=includeSCPI,
-            **kwargs
-        )
-
-    @staticmethod
-    def control(get_command, set_command, docs,
-                validator=lambda v, vs: v, values=(), map_values=False,
-                get_process=lambda v: v, set_process=lambda v: v,
-                check_set_errors=False, check_get_errors=False,
-                **kwargs):
-        """Fake Instrument.control.
-
-        Strip commands and only store and return values indicated by
-        format strings to mimic many simple commands.
-        This is analogous how the tests in test_instrument are handled.
-        """
-
-        # Regex search to find first format specifier in the command
-        fmt_spec_pattern = r'(%[\w.#-+ *]*[diouxXeEfFgGcrsa%])'
-        match = re.findall(fmt_spec_pattern, set_command)
-        if match:
-            # format_specifier = match.group(0)
-            format_specifier = ','.join(match)
-        else:
-            format_specifier = ''
-        # To preserve as much functionality as possible, call the real
-        # control method with modified get_command and set_command.
-        return Instrument.control(get_command="",
-                                  set_command=format_specifier,
-                                  docs=docs,
-                                  validator=validator,
-                                  values=values,
-                                  map_values=map_values,
-                                  get_process=get_process,
-                                  set_process=set_process,
-                                  check_set_errors=check_set_errors,
-                                  check_get_errors=check_get_errors,
-                                  **kwargs)
